Rename RawConfiguration to SchemaConfigurationFile
