+++ src/Configuration/Schema/SchemaConfigurationFile.php
@@ -33,54 +33,77 @@
 
 declare(strict_types=1);
 
-namespace Infection\Configuration\RawConfiguration;
+namespace Infection\Configuration\Schema;
 
-use function sprintf;
-use Throwable;
-use UnexpectedValueException;
+use Seld\JsonLint\JsonParser;
+use Seld\JsonLint\ParsingException;
+use stdClass;
+use function is_file;
+use function is_readable;
+use function Safe\file_get_contents;
 
 /**
  * @internal
  */
-final class InvalidFile extends UnexpectedValueException
+final class SchemaConfigurationFile
 {
-    public static function createForFileNotFound(RawConfiguration $config): self
+    private $path;
+
+    /**
+     * @var stdClass|null
+     */
+    private $decodedContents;
+
+    public function __construct(string $path)
     {
-        return new self(sprintf(
-            'The file "%s" could not be found or is not a file.',
-            $config->getPath()
-        ));
+        $this->path = $path;
     }
 
-    public static function createForFileNotReadable(RawConfiguration $config): self
+    public function getPath(): string
     {
-        return new self(sprintf(
-            'The file "%s" is not readable.',
-            $config->getPath()
-        ));
+        return $this->path;
     }
 
-    public static function createForCouldNotRetrieveFileContents(RawConfiguration $config): self
+    /**
+     * @throws InvalidFile
+     */
+    public function getDecodedContents(): stdClass
     {
-        return new self(sprintf(
-            'Could not retrieve the contents of the file "%s".',
-            $config->getPath()
-        ));
+        $this->initDecodedContents();
+
+        return $this->decodedContents;
     }
 
-    public static function createForInvalidJson(
-        RawConfiguration $config,
-        string $error,
-        Throwable $previous
-    ): self {
-        return new self(
-            sprintf(
-                'Could not parse the JSON file "%s": %s',
-                $config->getPath(),
-                $error
-            ),
-            0,
-            $previous
-        );
+    /**
+     * @throws InvalidFile
+     */
+    private function initDecodedContents(): void
+    {
+        if (null !== $this->decodedContents) {
+            return;
+        }
+
+        if (!is_file($this->path)) {
+            throw InvalidFile::createForFileNotFound($this);
+        }
+
+        if (!is_readable($this->path)) {
+            throw InvalidFile::createForFileNotReadable($this);
+        }
+
+        $contents = file_get_contents($this->path);
+
+        if (false === $contents) {
+            throw InvalidFile::createForCouldNotRetrieveFileContents($this);
+        }
+
+        try {
+            $this->decodedContents = (new JsonParser())->parse(
+                $contents,
+                JsonParser::DETECT_KEY_CONFLICTS
+            );
+        } catch (ParsingException $exception) {
+            throw InvalidFile::createForInvalidJson($this, $exception->getMessage(), $exception);
+        }
     }
 }
